mlops_templates/classification_template/{{ cookiecutter.project_name }}/requirements.txt
mlflow==2.9.2
databricks-registry-webhooks==0.1.2
PyYAML
ray[data,train,tune,serve]==2.9.3
joblib
tensorboardx